Rethink the idempotence a bit.
This uses the output from ruby-build --definitions to get the actual
name to install. It also means that the builds/ folders are named
after resource.name instead, and so might be duplicated when not needed.
This adds a VERSION file to track what actual version was installed
for a given resource.